+++ rand_distr/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+- New `Zeta` and `Zipf` distributions (#1136)
+
 ## [0.4.1] - 2021-06-15
 - Empirically test PDF of normal distribution (#1121)
 - Correctly document `no_std` support (#1100)