+++ rand_distr/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.2] - 2021-09-18
 - New `Zeta` and `Zipf` distributions (#1136)
 - New `SkewNormal` distribution (#1149)
 - New `Gumbel` and `Frechet` distributions (#1168, #1171)